+++ src/api/mod.rs
@@ -2,9 +2,13 @@
 #![allow(clippy::needless_lifetimes)]
 
 pub use adapter::{
-    local::LocalMockServerAdapter, standalone::RemoteMockServerAdapter, Method, MockServerAdapter,
+    local::LocalMockServerAdapter, Method, MockServerAdapter,
     Regex,
 };
+
+#[cfg(feature = "remote")]
+pub use adapter::{standalone::RemoteMockServerAdapter};
+
 pub use mock::{Mock, MockExt};
 pub use server::MockServer;
 pub use spec::{Then, When};