+++ tests/examples/mod.rs
@@ -7,10 +7,13 @@ mod file_body_tests;
 mod getting_started_tests;
 mod headers_tests;
 mod json_body_tests;
-mod multiserver_tests;
+mod multi_server_tests;
 mod query_param_tests;
 mod showcase_tests;
-mod standalone_tests;
 mod string_body_tests;
 mod url_matching_tests;
+
+#[cfg(feature = "remote")]
 mod x_www_form_urlencoded_tests;
+#[cfg(feature = "remote")]
+mod standalone_tests;